fixed g++4.3 compilation
git-svn-id: https://piggledy.org/svn/peerfuse/trunk@750 f7b84f85-2e04-4de5-9e99-c4612dc0539c

+++ common/cache_base.cpp
@@ -39,7 +39,7 @@ void CacheInterface::Write(std::string path, const char* buf, size_t size, off_t
 
 int CacheInterface::Read(std::string path, char* buf, size_t size, off_t off)
 {
-	size_t file_size = GetAttr(path).size;
+	size_t file_size = (size_t)GetAttr(path).size;
 
 	if(off > (off_t)file_size)
 		return 0;
